Preserving file permissions when extracting zip file

+++ nw/getnwfromnet.py
@@ -10,6 +10,28 @@ from nw import NWTMPDIR, is_win, is_darwin, is_linux, s3_node_webkit_url
 from nw.download import DownloadFile
 
 
+def GetPermission(zip, dst):
+  # build the destination pathname, replacing
+  # forward slashes to platform specific separators.
+  # Strip trailing path separator, unless it represents the root.
+  if (dst[-1:] in (os.path.sep, os.path.altsep)
+      and len(os.path.splitdrive(dst)[1]) > 1):
+    dst = dst[:-1]
+
+  members = zip.infolist()
+  for member in members:
+    targetpath = dst
+    # don't include leading "/" from file name if present
+    if member.filename[0] == '/':
+      targetpath = os.path.join(targetpath, member.filename[1:])
+    else:
+      targetpath = os.path.join(targetpath, member.filename)
+
+    targetpath = os.path.normpath(targetpath)
+    if os.path.isfile(targetpath):
+      mode = member.external_attr >> 16 & 0x1FF
+      os.chmod(targetpath, mode)
+
 def GetUrl(nw_name):
   base_url = s3_node_webkit_url
   f = urllib2.urlopen(base_url)
@@ -21,7 +43,7 @@ def GetUrl(nw_name):
     if child.tag.find('Contents') != -1 and child[0].text.find(nw_name) != -1:
       return base_url + child[0].text
 
-def GetNwFromNet(ver, target):
+def GetNwFromNet(ver, target, keep=False):
   path_ = NWTMPDIR
   if not os.path.isdir(path_):
     os.mkdir(path_)
@@ -57,10 +79,10 @@ def GetNwFromNet(ver, target):
     zip = zipfile.ZipFile(nw_tar_path_, 'r')
     file_name = os.path.dirname(zip.namelist()[0])
     nw_path_ = os.path.join(update_path_, file_name)
-    print nw_path_
     if os.path.isdir(nw_path_):
       shutil.rmtree(nw_path_)
     zip.extractall(update_path_)
+    GetPermission(zip, update_path_)
     zip.close()
 
   if target_platfrom == nwfiles.PLATFORMNAMEMAC:
@@ -68,9 +90,11 @@ def GetNwFromNet(ver, target):
     if os.path.isdir(nw_path_):
       shutil.rmtree(nw_path_)
     zip.extractall(nw_path_)
+    GetPermission(zip, nw_path_)
     zip.close()
 
-  # delete compress file
-  os.remove(nw_tar_path_)
+  if not keep:
+    # delete compress file
+    os.remove(nw_tar_path_)
 
   return nw_path_
